+++ client/src/app/app.tsx
@@ -13,7 +13,7 @@ import Game from './game'
 import Login from './login'
 import MainMenu from './main-menu'
 import BossLanding from './main-menu/boss-landing'
-import CreatePrivate from './main-menu/create-private'
+import PrivateGame from './main-menu/private-game'
 import Credits from './main-menu/credits'
 import DataSettings from './main-menu/data-settings'
 import GameSettings from './main-menu/game-settings'
@@ -55,8 +55,8 @@ function App() {
 					return <DataSettings setMenuSection={setMenuSection} />
 				case 'credits':
 					return <Credits setMenuSection={setMenuSection} />
-				case 'create-private':
-					return <CreatePrivate setMenuSection={setMenuSection} />
+				case 'private-game':
+					return <PrivateGame setMenuSection={setMenuSection} />
 				case 'boss-landing':
 					return <BossLanding setMenuSection={setMenuSection} />
 				case 'mainmenu':
